pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hrs"
version = "1.0.0"
description = "Health Recommendation System - A Flask-based web application providing AI-powered health recommendations"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "HRS Contributors"}
]
keywords = ["health", "recommendations", "ai", "llm", "rag", "flask", "medical"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Flask",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "pyyaml",
    "pandas==2.3.3",
    "protobuf",
    "Flask==3.1.2",
    "gunicorn==23.0.0",
    "openai==2.9.0",
    "python-dotenv==1.2.1",
    "google-generativeai==0.8.5",
    "chromadb==1.3.5",
    "sentence-transformers==5.1.2",
    "pyarrow",
    "numpy==2.3.3",
    "matplotlib==3.10.7",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "torch==2.8.0",
    "stable-baselines3==2.7.0",
    "torchvision==0.23.0",
    "transformers==4.57.0",
    "ucimlrepo==0.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/ekazum/HRS"
Repository = "https://github.com/ekazum/HRS"
Issues = "https://github.com/ekazum/HRS/issues"

[tool.setuptools]
packages = ["src.LLM", "src.RAG", "src.EHR", "src.P_CAFE"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | chromadb_data
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
